requirements.txt
pyTelegramBotAPI==4.14.0
python-dotenv==1.0.0
requests==2.31.0
urllib3==2.0.7
openai>=1.0.0
